requirements.txt
Django>=5.0,<6.0
django-environ>=0.11,<1.0
django-filter>=24.2,<25.0
mysqlclient>=2.2,<3.0
gunicorn>=22.0,<23.0
